+++ checkpoint/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Return empty dict (instead of None) as metadata for *empty* Mapping typed
+ unregistered types.
+
 
 ## [0.6.3] - 2024-09-10